+++ CHANGELOG.md
@@ -41,6 +41,11 @@ The current file describes significant changes in the code that may affect the u
 * Added *MultiImage* and *MultiField* field types to Entity Manager. Read more in [documentation](./src/Oro/Bundle/AttachmentBundle/README.md).
 
 ### Removed
+* `*.class` parameters for all entities were removed from the dependency injection container.
+The entity class names should be used directly, e.g. `'Oro\Bundle\EmailBundle\Entity\Email'`
+instead of `'%oro_email.email.entity.class%'` (in service definitions, datagrid config files, placeholders, etc.), and
+`\Oro\Bundle\EmailBundle\Entity\Email::class` instead of `$container->getParameter('oro_email.email.entity.class')`
+(in PHP code).
 
 #### ActivityListBundle
 * The `getActivityClass()` method was removed from `Oro\Bundle\ActivityListBundle\Model\ActivityListProviderInterface`.
@@ -183,7 +188,7 @@ The current file describes significant changes in the code that may affect the u
 ### Removed
 
 #### All Bundles
-* All `*.class` parameters were removed from the dependency injection container.
+* All `*.class` parameters for service definitions were removed from the dependency injection container.
 
 #### Math component
 * The deprecated method `Oro\Component\Math\BigDecimal::withScale()` was removed. Use `toScale()` method instead.